+++ cmd/gomodule-proxy/BUILD.bazel
@@ -1,0 +1,26 @@
+load("@rules_go//go:def.bzl", "go_binary", "go_library")
+
+go_library(
+    name = "gomodule-proxy_lib",
+    srcs = [
+        "command.go",
+        "main.go",
+    ],
+    importpath = "go.f110.dev/gomodule-proxy/cmd/gomodule-proxy",
+    visibility = ["//visibility:private"],
+    deps = [
+        "//cmd/gomodule-proxy/internal/config",
+        "//internal/gomodule",
+        "@com_github_google_go_github_v40//github:go_default_library",
+        "@com_github_spf13_cobra//:go_default_library",
+        "@com_github_spf13_pflag//:go_default_library",
+        "@org_golang_x_oauth2//:go_default_library",
+        "@org_golang_x_xerrors//:go_default_library",
+    ],
+)
+
+go_binary(
+    name = "gomodule-proxy",
+    embed = [":gomodule-proxy_lib"],
+    visibility = ["//visibility:public"],
+)
