+++ internal/gomodule/BUILD.bazel
@@ -1,0 +1,37 @@
+load("@rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "gomodule",
+    srcs = [
+        "fetcher.go",
+        "proxy.go",
+        "server.go",
+    ],
+    importpath = "go.f110.dev/gomodule-proxy/internal/gomodule",
+    visibility = ["//:__subpackages__"],
+    deps = [
+        "@com_github_go_git_go_git_v5//:go_default_library",
+        "@com_github_go_git_go_git_v5//plumbing:go_default_library",
+        "@com_github_go_git_go_git_v5//plumbing/filemode:go_default_library",
+        "@com_github_go_git_go_git_v5//plumbing/object:go_default_library",
+        "@com_github_google_go_github_v40//github:go_default_library",
+        "@com_github_gorilla_mux//:go_default_library",
+        "@org_golang_x_mod//modfile",
+        "@org_golang_x_mod//semver",
+        "@org_golang_x_tools_go_vcs//:vcs",
+        "@org_golang_x_xerrors//:go_default_library",
+    ],
+)
+
+go_test(
+    name = "gomodule_test",
+    srcs = ["fetcher_test.go"],
+    embed = [":gomodule"],
+    deps = [
+        "@com_github_go_git_go_git_v5//:go_default_library",
+        "@com_github_go_git_go_git_v5//plumbing/object:go_default_library",
+        "@com_github_stretchr_testify//assert:go_default_library",
+        "@com_github_stretchr_testify//require:go_default_library",
+        "@org_golang_x_tools_go_vcs//:vcs",
+    ],
+)
