+++ internal/gomodule/server.go
@@ -1,4 +1,4 @@
-package main
+package gomodule
 
 import (
 	"context"
